move to package com.powsybl.dynawaltz.dsl
