+++ dynawaltz-dsl/src/main/groovy/com/powsybl/dynawaltz/dsl/dynamicmodels/GeneratorModelGroovyExtension.groovy
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.dynawaltz.dynamicmodels
+package com.powsybl.dynawaltz.dsl.dynamicmodels
 
 import java.util.function.Consumer
