+++ dynawaltz-dsl/src/test/java/com/powsybl/dynawaltz/dsl/DynaWaltzGroovyEventModelsSupplierTest.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.dynawaltz;
+package com.powsybl.dynawaltz.dsl;
 
 import com.google.common.jimfs.Configuration;
 import com.google.common.jimfs.Jimfs;
@@ -13,10 +13,11 @@ import com.powsybl.dynamicsimulation.EventModelsSupplier;
 import com.powsybl.dynamicsimulation.groovy.EventModelGroovyExtension;
 import com.powsybl.dynamicsimulation.groovy.GroovyEventModelsSupplier;
 import com.powsybl.dynamicsimulation.groovy.GroovyExtension;
+import com.powsybl.dynawaltz.DynaWaltzProvider;
 import com.powsybl.dynawaltz.events.AbstractBlackBoxEventModel;
 import com.powsybl.dynawaltz.events.EventQuadripoleDisconnection;
-import com.powsybl.dynawaltz.events.EventQuadripoleDisconnectionGroovyExtension;
-import com.powsybl.dynawaltz.events.EventSetPointBooleanGroovyExtension;
+import com.powsybl.dynawaltz.dsl.events.EventQuadripoleDisconnectionGroovyExtension;
+import com.powsybl.dynawaltz.dsl.events.EventSetPointBooleanGroovyExtension;
 import com.powsybl.iidm.network.Identifiable;
 import com.powsybl.iidm.network.Line;
 import com.powsybl.iidm.network.Network;
