+++ dynawaltz-dsl/src/test/java/com/powsybl/dynawaltz/dsl/ieee/AbstractDynaWaltzLocalCommandExecutor.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.dynawaltz.ieee;
+package com.powsybl.dynawaltz.dsl.ieee;
 
 import static org.junit.Assert.assertFalse;
