+++ dynawaltz-dsl/src/test/java/com/powsybl/dynawaltz/dsl/ieee57/Ieee57DisconnectGeneratorLocalCommandExecutor.java
@@ -4,10 +4,10 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.dynawaltz.ieee.ieee57;
+package com.powsybl.dynawaltz.dsl.ieee57;
 
 import com.powsybl.dynawaltz.DynaWaltzParameters;
-import com.powsybl.dynawaltz.ieee.AbstractDynaWaltzLocalCommandExecutor;
+import com.powsybl.dynawaltz.dsl.ieee.AbstractDynaWaltzLocalCommandExecutor;
 import com.powsybl.iidm.network.Network;
 
 import static com.powsybl.dynawaltz.xml.DynaWaltzConstants.*;
